+++ packages/react-charts/src/components/ChartCursorTooltip/ChartCursorTooltip.tsx
@@ -32,7 +32,7 @@ export interface ChartCursorTooltipProps extends ChartTooltipProps {
   /**
    * The angle prop specifies the angle to rotate the tooltip around its origin point.
    */
-  angle?: string | number;
+  angle?: number;
   /**
    * The center prop determines the position of the center of the tooltip flyout. This prop should be given as an object
    * that describes the desired x and y svg coordinates of the center of the tooltip. This prop is useful for
@@ -226,7 +226,7 @@ export interface ChartCursorTooltipProps extends ChartTooltipProps {
    *
    * @propType number | string | Function | string[] | number[]
    */
-  text?: StringOrNumberOrCallback | string[] | number[];
+  text?: string[] | StringOrNumberOrCallback;
   /**
    * The theme prop specifies a theme to use for determining styles and layout properties for a component. Any styles or
    * props defined in theme may be overwritten by props specified on the component instance.