+++ packages/react-core/src/next/components/DualListSelector/__tests__/DualListSelector.test.tsx
@@ -1,0 +1,26 @@
+import { render } from '@testing-library/react';
+import { DualListSelectorPane } from '../../DualListSelector';
+import { SearchInput } from '../../../../components/SearchInput';
+import React from 'react';
+
+describe('DualListSelector', () => {
+  test('basic', () => {
+    const { asFragment } = render(<DualListSelectorPane id="basicTest" />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  test('with search inputs', () => {
+    const { asFragment } = render(<DualListSelectorPane id="searchTest" searchInput={<SearchInput />} />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  test('with custom status', () => {
+    const { asFragment } = render(<DualListSelectorPane id="statusTest" status="Test status1" />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  test('basic with disabled controls', () => {
+    const { asFragment } = render(<DualListSelectorPane id="disabledTest" isDisabled />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});