+++ packages/react-core/src/next/components/index.ts
@@ -1,1 +1,2 @@
 export * from './Modal';
+export * from './DualListSelector';