+++ packages/react-table/src/demos/examples/TableFilterable.tsx
@@ -6,6 +6,8 @@ import {
   EmptyState,
   EmptyStateActions,
   EmptyStateBody,
+  EmptyStateIcon,
+  EmptyStateHeader,
   EmptyStateFooter,
   Label,
   MenuToggle,
@@ -306,7 +308,12 @@ export const TableFilterable: React.FunctionComponent = () => {
       : rowData;
 
   const emptyState = (
-    <EmptyState headingLevel="h5" titleText="Clear all filters and try again." icon={SearchIcon}>
+    <EmptyState>
+      <EmptyStateHeader
+        titleText="Clear all filters and try again."
+        headingLevel="h5"
+        icon={<EmptyStateIcon icon={SearchIcon} />}
+      />
       <EmptyStateBody>No results match the filter criteria. Clear all filters and try again.</EmptyStateBody>
       <EmptyStateFooter>
         <EmptyStateActions>