+++ packages/react-charts/CHANGELOG.md
@@ -113,6 +113,50 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @patternfly/react-charts
 
+# 7.3.0-prerelease.3 (2024-02-28)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.3.0-prerelease.2 (2024-02-28)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.3.0-prerelease.1 (2024-02-14)
+
+### Bug Fixes
+
+- **charts:** adjust for Victory types changes ([#10082](https://github.com/patternfly/patternfly-react/issues/10082)) ([714eaca](https://github.com/patternfly/patternfly-react/commit/714eaca02f99f709a6aadc1d9dd58d25940f7403))
+
+# 7.3.0-prerelease.0 (2024-02-02)
+
+# 5.2.0 (2024-02-02)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0 (2024-02-02)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0-prerelease.10 (2024-01-15)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0-prerelease.9 (2024-01-11)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0-prerelease.8 (2023-12-11)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0-prerelease.7 (2023-11-10)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.2.0-prerelease.6 (2023-11-03)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
 # 7.2.0-prerelease.5 (2023-10-26)
 
 **Note:** Version bump only for package @patternfly/react-charts