+++ packages/react-tokens/CHANGELOG.md
@@ -105,6 +105,34 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @patternfly/react-tokens
 
+# 5.3.0-prerelease.1 (2024-02-28)
+
+### Bug Fixes
+
+- **deps:** bump to latest core ([#10116](https://github.com/patternfly/patternfly-react/issues/10116)) ([dc1aceb](https://github.com/patternfly/patternfly-react/commit/dc1aceb0ee880631320b85bfb068be13d7852a37))
+
+# 5.3.0-prerelease.0 (2024-02-02)
+
+# 5.2.0 (2024-02-02)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+# 5.2.0 (2024-02-02)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+# 5.2.0-prerelease.7 (2023-12-11)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+# 5.2.0-prerelease.6 (2023-11-10)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
+# 5.2.0-prerelease.5 (2023-11-03)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
 # 5.2.0-prerelease.4 (2023-10-26)
 
 **Note:** Version bump only for package @patternfly/react-tokens